requirements.txt
python-dotenv>=0.19.2
cryptography>=35.0.0